+++ rbkApiModules.Testing.Core.Tests/Usings.cs
@@ -1,0 +1,1 @@
+global using rbkApiModules.Commons.Testing;